Move ResolveRef test to new package
